Create Calc-Euler's Method.py
This is a program that is able to run the Euler's Method function. Euler's Method is a calculus algorithm that is used to approximate your solution by increasing the value of your input by a certain value until your input is the input you want to find the solution of.